+++ settings.lua
@@ -12,6 +12,9 @@ local function get_setting(setting, default, type)
         better_commands.settings[setting] = minetest.settings:get_bool(long_setting, default)
     elseif type == "number" then
         better_commands.settings[setting] = tonumber(minetest.settings:get(long_setting)) or default
+    elseif type == "comma_separated" then
+        local value = minetest.settings:get(long_setting)
+        better_commands.settings[setting] = value and value:split(",") or default
     end
 end
 
@@ -20,6 +23,7 @@ local settings = {
     {"acovg_time", false, "bool"},
     {"save_interval", 3, "number"},
     {"kill_creative_players", false, "bool"},
+    {"clear_lists", {"main", "craft", "offhand"}, "comma_separated"},
 
     {"scoreboard_picked_up", true, "bool"},
     {"scoreboard_mined", true, "bool"},